+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 02-23-2020
+### Changed
+- Removed support for Breadcrumbs NavXT
+- Add support to Yoast SEO Breadcrumbs
+- Fix: Site footer width on mobile
+
 ## [1.0.3] - 02-23-2020
 ### Changed
 - The released version of theme doesn't come with sass files anymore